Create abstract signal class for plugins
In order to create an abstract class for signal handling, it was necessary to
remove the signals methods from the apps of the plugins, and put the
implementation from the signal on ColabProxiedAppConfig. It was also created a
abstract class on plugins utils that provide abstract methods to register and
connect signals.

Signed-off-by: Lucas Kanashiro <kanashiro.duarte@gmail.com>